moving interactive clients to own files
