+++ ui/v2.5/src/hooks/Interactive/interactive.ts
@@ -1,29 +1,16 @@
-import Handy from "thehandy";
-import {
-  HandyMode,
-  HsspSetupResult,
-  CsvUploadResponse,
-  HandyFirmwareStatus,
-} from "thehandy/lib/types";
-import {
-  ButtplugClient,
-  ButtplugClientDevice,
-  ButtplugBrowserWebsocketClientConnector
-} from "buttplug";
-
-interface IFunscript {
+export interface IFunscript {
   actions: Array<IAction>;
   inverted: boolean;
   range: number;
 }
 
-interface IAction {
+export interface IAction {
   at: number;
   pos: number;
 }
 
 // Utility function to convert one range of values to another
-function convertRange(
+export function convertRange(
   value: number,
   fromLow: number,
   fromHigh: number,
@@ -36,7 +23,7 @@ function convertRange(
 // Converting to CSV first instead of uploading Funscripts is required
 // Reference for Funscript format:
 // https://pkg.go.dev/github.com/funjack/launchcontrol/protocol/funscript
-function convertFunscriptToCSV(funscript: IFunscript) {
+export function convertFunscriptToCSV(funscript: IFunscript) {
   const lineTerminator = "\r\n";
   if (funscript?.actions?.length > 0) {
     return funscript.actions.reduce((prev: string, curr: IAction) => {
@@ -59,42 +46,6 @@ function convertFunscriptToCSV(funscript: IFunscript) {
   throw new Error("Not a valid funscript");
 }
 
-// copied from https://github.com/defucilis/thehandy/blob/main/src/HandyUtils.ts
-// since HandyUtils is not exported.
-// License is listed as MIT. No copyright notice is provided in original.
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-async function uploadCsv(
-  csv: File,
-  filename?: string
-): Promise<CsvUploadResponse> {
-  const url = "https://www.handyfeeling.com/api/sync/upload?local=true";
-  if (!filename) filename = "script_" + new Date().valueOf() + ".csv";
-  const formData = new FormData();
-  formData.append("syncFile", csv, filename);
-  const response = await fetch(url, {
-    method: "post",
-    body: formData,
-  });
-  const newUrl = await response.json();
-  return newUrl;
-}
-
 export interface IInteractive {
   scriptOffset: number;
   enabled(): boolean;
@@ -106,187 +57,4 @@ export interface IInteractive {
   pause(): Promise<void>;
   ensurePlaying(position: number): Promise<void>;
   setLooping(looping: boolean): Promise<void>;
-}
-
-// Interactive currently uses the Handy API, but could be expanded to use buttplug.io
-// via buttplugio/buttplug-rs-ffi's WASM module.
-export class HandyInteractive implements IInteractive {
-  _connected: boolean;
-  _playing: boolean;
-  _scriptOffset: number;
-  _handy: Handy;
-
-  constructor(scriptOffset: number = 0) {
-    this._handy = new Handy();
-    this._scriptOffset = scriptOffset;
-    this._connected = false;
-    this._playing = false;
-  }
-
-  enabled(): boolean {
-    return (this._handy.connectionKey !== "");
-  }
-
-  async connect() {
-    const connected = await this._handy.getConnected();
-    if (!connected) {
-      throw new Error("Handy not connected");
-    }
-
-    // check the firmware and make sure it's compatible
-    const info = await this._handy.getInfo();
-    if (info.fwStatus === HandyFirmwareStatus.updateRequired) {
-      throw new Error("Handy firmware update required");
-    }
-  }
-
-  set handyKey(key: string) {
-    this._handy.connectionKey = key;
-  }
-
-  get handyKey(): string {
-    return this._handy.connectionKey;
-  }
-
-  set scriptOffset(offset: number) {
-    this._scriptOffset = offset;
-  }
-
-  async uploadScript(funscriptPath: string) {
-    if (!(this._handy.connectionKey && funscriptPath)) {
-      return;
-    }
-
-    const csv = await fetch(funscriptPath)
-      .then((response) => response.json())
-      .then((json) => convertFunscriptToCSV(json));
-    const fileName = `${Math.round(Math.random() * 100000000)}.csv`;
-    const csvFile = new File([csv], fileName);
-
-    const tempURL = await uploadCsv(csvFile).then((response) => response.url);
-
-    await this._handy.setMode(HandyMode.hssp);
-
-    this._connected = await this._handy
-      .setHsspSetup(tempURL)
-      .then((result) => result === HsspSetupResult.downloaded);
-  }
-
-  async sync() {
-    return this._handy.getServerTimeOffset();
-  }
-
-  setServerTimeOffset(offset: number) {
-    this._handy.estimatedServerTimeOffset = offset;
-  }
-
-  async play(position: number) {
-    if (!this._connected) {
-      return;
-    }
-
-    this._playing = await this._handy
-      .setHsspPlay(
-        Math.round(position * 1000 + this._scriptOffset),
-        this._handy.estimatedServerTimeOffset + Date.now() // our guess of the Handy server's UNIX epoch time
-      )
-      .then(() => true);
-  }
-
-  async pause() {
-    if (!this._connected) {
-      return;
-    }
-    this._playing = await this._handy.setHsspStop().then(() => false);
-  }
-
-  async ensurePlaying(position: number) {
-    if (this._playing) {
-      return;
-    }
-    await this.play(position);
-  }
-
-  async setLooping(looping: boolean) {
-    if (!this._connected) {
-      return;
-    }
-    this._handy.setHsspLoop(looping);
-  }
-}
-
-export class ButtplugInteractive implements IInteractive {
-  _scriptOffset: number;
-
-  constructor(scriptOffset: number = 0) {
-    this._scriptOffset = scriptOffset;
-  }
-
-  enabled(): boolean {
-    return true;
-  }
-
-  async connect() {
-    const connector = new ButtplugBrowserWebsocketClientConnector("ws://localhost:12345");
-    const client = new ButtplugClient("Device Control Example");
-    client.addListener(
-      "deviceadded",
-      async (device: ButtplugClientDevice) => {
-        console.log(`Device Connected: ${device.name}`);
-        //devices.current.push(device);
-        // setDeviceDatas((deviceDatas) => [
-        //   ...deviceDatas,
-        //   {
-        //     intensities: { vibration: 0, rotation: 0 },
-        //   },
-        // ]);
-      }
-    );
-    client.addListener("deviceremoved", (device) =>
-      console.log(`Device Removed: ${device.name}`)
-    );
-    await client.connect(connector);
-    await client.startScanning();
-  }
-
-  set scriptOffset(offset: number) {
-    this._scriptOffset = offset;
-  }
-
-  async uploadScript(funscriptPath: string) {
-    if (!funscriptPath) {
-      return;
-    }
-
-    const csv = await fetch(funscriptPath)
-      .then((response) => response.json())
-      .then((json) => convertFunscriptToCSV(json));
-    const fileName = `${Math.round(Math.random() * 100000000)}.csv`;
-    const csvFile = new File([csv], fileName);
-    return;
-  }
-
-  async sync() {
-    return 0;
-  }
-
-  setServerTimeOffset(offset: number) {
-    return;
-  }
-
-  async play(position: number) {
-    return;
-  }
-
-  async pause() {
-    return;
-  }
-
-  async ensurePlaying(position: number) {
-    return;
-  }
-
-  async setLooping(looping: boolean) {
-    return;
-  }
 }
